chore: Move Filter components to "common" folder
