--- src/components/common/Button.tsx
@@ -1,8 +1,0 @@
-export default function Button() {
-  return (
-    <>
-      <p>This is Empty Component</p>
-      <p>Make your own component</p>
-    </>
-  );
-}
