--- src/components/common/Card.tsx
@@ -1,8 +1,0 @@
-export default function Card() {
-  return (
-    <>
-      <p>This is Empty Component</p>
-      <p>Make your own component</p>
-    </>
-  );
-}
