+++ src/components/common/ResultCount.tsx
