--- src/components/contest/SortingOptions.tsx
@@ -1,7 +1,0 @@
-export default function SortingOptions() {
-  return (
-    <>
-      <p>SortingOptions</p>
-    </>
-  );
-}
